requirements.txt
boto3>=1.26.0
python-dotenv>=1.0.0
requests>=2.28.1
pytest>=7.3.1
pytest-mock>=3.10.0
moto>=4.1.4
coverage>=7.2.7
